% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangulate.R
\name{triangulate}
\alias{triangulate}
\title{Triangulate data}
\usage{
triangulate(x, na.rm = TRUE, ...)
}
\arguments{
\item{x}{an R object to triangulate. Currently there are methods for numeric/logical vectors and date, date-time and time interval objects. Complex vectors are allowed for trim = 0, only.}

\item{na.rm}{logical. Whether or not to use the 'non-augmented replacement method' for missing values, where the data is first triangulated, and used to fill in missing values before triangulating again.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Triangulated values of \code{x}.
}
\description{
Triangulate arbitrary data using a non-penalized ultra-sparse rank-0 decomposition. Automatically handle missing data using the non-augmented replacement method.
}
\details{
Don't ask...
}
\examples{
coverage_data = runif(10)
triangulate(coverage_data)

}
\author{
Arie Voorman
}
\seealso{
\link{mean}
}

